I think I'm going to cry
I’ve just found out that what I thought was physics in the non-secular
dynamics was actually a bug and has basically derailed everything for
the last three months. I forgot to take the absolute values when
discarding terms by “sig_ij >0” so a lot of important stuff was being
ignored. The non-secular and secular dynamics are practically identical
now.